+++ CHANGES.md
@@ -36,6 +36,8 @@
 - Support `(foreign_objects ...)` field in `(executable ...)` and `(library
   ...)` stanzas (#6084, fixes #4129, @gridbugs)
 
+- Fix compilation of Dune under esy on Windows (#6109, fixes #6098, @nojb)
+
 3.4.1 (26-07-2022)
 ------------------